Restructured include directory
