+++ src/adaptations/weave-device/EchoServer.cpp
@@ -16,8 +16,8 @@
  *    limitations under the License.
  */
 
-#include <internal/WeaveDeviceInternal.h>
-#include <internal/EchoServer.h>
+#include <Weave/DeviceLayer/internal/WeaveDeviceInternal.h>
+#include <Weave/DeviceLayer/internal/EchoServer.h>
 
 using namespace ::nl;
 using namespace ::nl::Weave;
