+++ src/adaptations/weave-device/Entropy.cpp
@@ -16,7 +16,7 @@
  *    limitations under the License.
  */
 
-#include <internal/WeaveDeviceInternal.h>
+#include <Weave/DeviceLayer/internal/WeaveDeviceInternal.h>
 #include <Weave/Support/crypto/WeaveRNG.h>
 
 using namespace ::nl;
