+++ src/adaptations/weave-device/FabricProvisioningServer.cpp
@@ -16,8 +16,8 @@
  *    limitations under the License.
  */
 
-#include <internal/WeaveDeviceInternal.h>
-#include <internal/FabricProvisioningServer.h>
+#include <Weave/DeviceLayer/internal/WeaveDeviceInternal.h>
+#include <Weave/DeviceLayer/internal/FabricProvisioningServer.h>
 
 using namespace ::nl;
 using namespace ::nl::Weave;
