+++ src/adaptations/weave-device/Logging.cpp
@@ -16,7 +16,7 @@
  *    limitations under the License.
  */
 
-#include <internal/WeaveDeviceInternal.h>
+#include <Weave/DeviceLayer/internal/WeaveDeviceInternal.h>
 #include <Weave/Support/logging/WeaveLogging.h>
 
 using namespace ::nl::Weave;
