+++ src/adaptations/weave-device/NetworkProvisioningServer.cpp
@@ -16,10 +16,10 @@
  *    limitations under the License.
  */
 
-#include <internal/WeaveDeviceInternal.h>
-#include <internal/NetworkProvisioningServer.h>
-#include <internal/NetworkInfo.h>
-#include <internal/ESPUtils.h>
+#include <Weave/DeviceLayer/internal/WeaveDeviceInternal.h>
+#include <Weave/DeviceLayer/internal/NetworkProvisioningServer.h>
+#include <Weave/DeviceLayer/internal/NetworkInfo.h>
+#include <Weave/DeviceLayer/internal/ESPUtils.h>
 
 #include <Weave/Core/WeaveTLV.h>
 #include <Weave/Profiles/WeaveProfiles.h>
