+++ src/adaptations/weave-device/include/Weave/DeviceLayer/ConfigurationManager.h
