+++ src/adaptations/weave-device/include/Weave/DeviceLayer/TraitManager.h
