+++ src/adaptations/weave-device/include/Weave/DeviceLayer/WeaveDeviceError.h
