+++ src/adaptations/weave-device/include/Weave/DeviceLayer/internal/BLEManager.h
