+++ src/adaptations/weave-device/include/Weave/DeviceLayer/internal/DeviceDescriptionServer.h
@@ -16,40 +16,38 @@
  *    limitations under the License.
  */
 
-#ifndef FABRIC_PROVISIONING_SERVER_H
-#define FABRIC_PROVISIONING_SERVER_H
+#ifndef DEVICE_DESCRIPTION_SERVER_H
+#define DEVICE_DESCRIPTION_SERVER_H
 
-#include <internal/WeaveDeviceInternal.h>
-#include <Weave/Profiles/fabric-provisioning/FabricProvisioning.h>
+#include <Weave/DeviceLayer/internal/WeaveDeviceInternal.h>
+#include <Weave/Profiles/device-description/DeviceDescription.h>
 
 namespace nl {
 namespace Weave {
 namespace Device {
 namespace Internal {
 
-class FabricProvisioningServer
-        : public ::nl::Weave::Profiles::FabricProvisioning::FabricProvisioningServer,
-          public ::nl::Weave::Profiles::FabricProvisioning::FabricProvisioningDelegate
+class DeviceDescriptionServer : public ::nl::Weave::Profiles::DeviceDescription::DeviceDescriptionServer
 {
-public:
-    typedef ::nl::Weave::Profiles::FabricProvisioning::FabricProvisioningServer ServerBaseClass;
+    typedef ::nl::Weave::Profiles::DeviceDescription::DeviceDescriptionServer ServerBaseClass;
 
+public:
     WEAVE_ERROR Init();
-    virtual WEAVE_ERROR HandleCreateFabric(void);
-    virtual WEAVE_ERROR HandleJoinExistingFabric(void);
-    virtual WEAVE_ERROR HandleLeaveFabric(void);
-    virtual WEAVE_ERROR HandleGetFabricConfig(void);
-    WEAVE_ERROR LeaveFabric(void);
-    virtual bool IsPairedToAccount() const;
 
     void OnPlatformEvent(const WeaveDeviceEvent * event);
+
+private:
+    static void HandleIdentifyRequest(void *appState, uint64_t nodeId, const IPAddress& nodeAddr,
+            const ::nl::Weave::Profiles::DeviceDescription::IdentifyRequestMessage& reqMsg, bool& sendResp,
+            ::nl::Weave::Profiles::DeviceDescription::IdentifyResponseMessage& respMsg);
 };
 
-extern FabricProvisioningServer FabricProvisioningSvr;
+extern DeviceDescriptionServer DeviceDescriptionSvr;
 
 } // namespace Internal
 } // namespace Device
 } // namespace Weave
 } // namespace nl
 
-#endif // FABRIC_PROVISIONING_SERVER_H
+
+#endif // DEVICE_DESCRIPTION_SERVER_H
