+++ src/adaptations/weave-device/include/Weave/DeviceLayer/internal/EchoServer.h
@@ -19,7 +19,7 @@
 #ifndef ECHO_SERVER_H
 #define ECHO_SERVER_H
 
-#include <internal/WeaveDeviceInternal.h>
+#include <Weave/DeviceLayer/internal/WeaveDeviceInternal.h>
 #include <Weave/Profiles/echo/Next/WeaveEchoServer.h>
 
 namespace nl {
