+++ src/adaptations/weave-device/include/Weave/DeviceLayer/internal/NetworkInfo.h
@@ -19,7 +19,7 @@
 #ifndef NETWORK_INFO_H
 #define NETWORK_INFO_H
 
-#include <internal/WeaveDeviceInternal.h>
+#include <Weave/DeviceLayer/internal/WeaveDeviceInternal.h>
 #include <Weave/Core/WeaveTLV.h>
 #include <Weave/Profiles/network-provisioning/NetworkProvisioning.h>
