+++ src/adaptations/weave-device/include/Weave/DeviceLayer/internal/ServiceDirectoryManager.h
@@ -16,24 +16,28 @@
  *    limitations under the License.
  */
 
-#ifndef TUNNEL_AGENT_H
-#define TUNNEL_AGENT_H
+#ifndef SERVICE_DIRECTORY_MANAGER_H
+#define SERVICE_DIRECTORY_MANAGER_H
 
-#include <internal/WeaveDeviceInternal.h>
-#include <Weave/Profiles/weave-tunneling/WeaveTunnelAgent.h>
+#include <Weave/DeviceLayer/internal/WeaveDeviceInternal.h>
+#include <Weave/Profiles/service-directory/ServiceDirectory.h>
 
 namespace nl {
 namespace Weave {
 namespace Device {
 namespace Internal {
 
-extern ::nl::Weave::Profiles::WeaveTunnel::WeaveTunnelAgent ServiceTunnelAgent;
+#if WEAVE_CONFIG_ENABLE_SERVICE_DIRECTORY
 
-extern WEAVE_ERROR InitServiceTunnelAgent();
+extern ::nl::Weave::Profiles::ServiceDirectory::WeaveServiceManager ServiceDirectoryMgr;
+
+extern WEAVE_ERROR InitServiceDirectoryManager(void);
+
+#endif // WEAVE_CONFIG_ENABLE_SERVICE_DIRECTORY
 
 } // namespace Internal
 } // namespace Device
 } // namespace Weave
 } // namespace nl
 
-#endif // TUNNEL_AGENT_H
+#endif // SERVICE_DIRECTORY_MANAGER_H
