+++ src/adaptations/weave-device/include/Weave/DeviceLayer/internal/WeaveDeviceConfig-ESP32.h
